Move generateCookie
