+++ src/utilities/tenancy/generateCookie.ts
